Alter the examples in the readme to Before/After
Update the README file to have the examples in a Before -> After style.

Showing the default Jackson annotations before the AutoJackson annotations
is easier to read. People can look at what they are used to using, and then
compare it to using Autojackson; rather than trying to understand AutoJackson
first, and seeing what they are used to second.

[skip ci]