Adjust ImplementationCreator for generics
At the moment generics fail, as we do not look for the actual type of the
Generic.

see https://github.com/peckb1/autojackson/issues/20

This starts the process of fixing that bug by correctly swapping generic return
types for the methods that use those generics as return types.

Tests currently fail on this commit - but I want to clean up a little and need
to commit to save at least compilation progress before I lose it.